+++ src/api/schemas/queries/weather.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class WeatherParameters(BaseModel):
+    path: str
