--- src/crawler/abstract_crawler.py
@@ -1,7 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class AbstractCrawler(ABC):
-    @abstractmethod
-    def crawl_content(self):
-        pass
