+++ src/csv_getter/crawler/abstract_crawler.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class AbstractCrawler(ABC):
+    @abstractmethod
+    def crawl_content(self):
+        pass
