+++ src/csv_getter/scrapper/abstract_scrapper.py
