--- src/parser/abstract_parser.py
@@ -1,7 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class AbstractParser(ABC):
-    @abstractmethod
-    def parse_content(self, html, *args, **kwargs):
-        pass
